+++ example/src/test/resources/extensions/Collection.feature
@@ -1,5 +1,5 @@
-@withA
 Feature: How an array list functioned
+
     Scenario: when the collection is empty
     Given it has "abc"
     Then it contains "abc"
