+++ example/src/test/resources/extensions/withA.feature
@@ -1,10 +1,20 @@
-@withExtra
-Feature: How an array list functioned
+Feature: Testing that background is add at the first steps
 
       Background: testing background
             Given a thing
+            And a thing a different
 
-      Scenario: with extra Scenario for A
+      Scenario: with 1 extra Scenario for A
+            Given step 1
+            When step 2
+            Then step 3
+
+      Scenario: with 2 extra Scenario for A
+            Given step 1
+            When step 2
+            Then step 3
+
+      Scenario: with 3 extra Scenario for A
             Given step 1
             When step 2
             Then step 3
