+++ example/src/test/resources/extensions/withB.feature
@@ -1,4 +1,6 @@
-Feature: How an array list functioned
+@withExtra
+Feature: Test that it adds the sub features of the tags on the sub feature
+
   Scenario: with extra Scenario for B
     Given step 1
     When step 2
